# Speed up WireNet.mapChannelForPoint