@kitlith 2019-08-25
to preserve the insertion order yet have open be a set, we could use a [LinkedHashSet](https://www.scala-lang.org/api/current/scala/collection/mutable/LinkedHashSet.html)? Should be less memory than having a seperate queue and set, but a bit more than just a queue or a set.